third_party/WebKit/LayoutTests/accessibility/inert-node-is-disabled-expected.txt


This tests that inert elements are not enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textField.isEnabled is true
PASS textField.isEnabled is false
PASS successfullyParsed is true

TEST COMPLETE
